+++ packages/polaris-viz/src/components/LineChart/Chart.tsx
@@ -19,7 +19,11 @@ import type {
   YAxisOptions,
 } from '@shopify/polaris-viz-core';
 
-import type {RenderTooltipContentData} from '../../types';
+import {Annotations} from '../Annotations';
+import type {
+  RenderTooltipContentData,
+  AnnotationLookupTable,
+} from '../../types';
 import {useXAxisLabels} from '../../hooks/useXAxisLabels';
 import {LinearXAxisLabels} from '../LinearXAxisLabels';
 import {useLegend, LegendContainer} from '../LegendContainer';
@@ -43,6 +47,7 @@ import {
   LineChartMargin as Margin,
   XMLNS,
   LABEL_AREA_TOP_SPACING,
+  ANNOTATIONS_LABELS_OFFSET,
 } from '../../constants';
 import {VisuallyHiddenRows} from '../VisuallyHiddenRows';
 import {YAxis} from '../YAxis';
@@ -59,6 +64,7 @@ import {yAxisMinMax} from './utilities';
 export interface ChartProps {
   isAnimated: boolean;
   renderTooltipContent: (data: RenderTooltipContentData) => React.ReactNode;
+  annotationsLookupTable: AnnotationLookupTable;
   data: LineChartDataSeriesWithDefaults[];
   showLegend: boolean;
   xAxisOptions: Required<XAxisOptions>;
@@ -75,6 +81,7 @@ const TOOLTIP_POSITION = {
 
 export function Chart({
   data,
+  annotationsLookupTable,
   dimensions,
   renderTooltipContent,
   showLegend = true,
@@ -91,6 +98,7 @@ export function Chart({
   const [activeIndex, setActiveIndex] = useState<number | null>(null);
   const [activeLineIndex, setActiveLineIndex] = useState(-1);
   const [labelHeight, setLabelHeight] = useState(0);
+  const [annotationsHeight, setAnnotationsHeight] = useState(0);
 
   const {legend, setLegendDimensions, height, width} = useLegend({
     data: [
@@ -117,8 +125,13 @@ export function Chart({
   const emptyState =
     data.length === 0 || data.every((series) => series.data.length === 0);
 
+  const chartYPosition = (Margin.Top as number) + annotationsHeight;
+
   const drawableHeight =
-    height - labelHeight - LABEL_AREA_TOP_SPACING - Margin.Top;
+    height - labelHeight - LABEL_AREA_TOP_SPACING - chartYPosition;
+
+  const annotationsDrawableHeight =
+    chartYPosition + drawableHeight + ANNOTATIONS_LABELS_OFFSET;
 
   const {minY, maxY} = yAxisMinMax(data);
 
@@ -241,8 +254,10 @@ export function Chart({
     width,
     height,
     x: chartStartPosition,
-    y: Margin.Top,
+    y: chartYPosition,
   };
+
+  const hasAnnotations = Object.keys(annotationsLookupTable).length > 0;
 
   return (
     <div className={styles.Container} style={{width, height}}>
@@ -260,9 +275,7 @@ export function Chart({
           <LinearXAxisLabels
             chartHeight={height}
             chartX={chartStartPosition - xAxisDetails.labelWidth / 2}
-            chartY={
-              drawableHeight + LABEL_AREA_TOP_SPACING + (Margin.Top as number)
-            }
+            chartY={drawableHeight + LABEL_AREA_TOP_SPACING + chartYPosition}
             labels={labels}
             labelWidth={xAxisDetails.labelWidth}
             onHeightChange={setLabelHeight}
@@ -279,7 +292,7 @@ export function Chart({
             theme={theme}
             transform={{
               x: selectedTheme.grid.horizontalOverflow ? 0 : chartStartPosition,
-              y: Margin.Top,
+              y: chartYPosition,
             }}
             width={
               selectedTheme.grid.horizontalOverflow ? width : drawableWidth
@@ -287,7 +300,7 @@ export function Chart({
           />
         ) : null}
 
-        <g transform={`translate(0,${Margin.Top})`}>
+        <g transform={`translate(0,${chartYPosition})`}>
           <YAxis
             ticks={ticks}
             width={yAxisLabelWidth}
@@ -298,7 +311,7 @@ export function Chart({
         </g>
 
         {emptyState ? null : (
-          <g transform={`translate(${chartStartPosition},${Margin.Top})`}>
+          <g transform={`translate(${chartStartPosition},${chartYPosition})`}>
             <Crosshair
               x={getXPosition({isCrosshair: true})}
               height={drawableHeight}
@@ -316,7 +329,27 @@ export function Chart({
           />
         )}
 
-        <g transform={`translate(${chartStartPosition},${Margin.Top})`}>
+        {hasAnnotations && (
+          <g
+            transform={`translate(${
+              chartStartPosition - xAxisDetails.labelWidth / 2
+            },0)`}
+            tabIndex={-1}
+          >
+            <Annotations
+              annotationsLookupTable={annotationsLookupTable}
+              axisLabelWidth={xAxisDetails.labelWidth}
+              drawableHeight={annotationsDrawableHeight}
+              drawableWidth={drawableWidth}
+              labels={labels}
+              onHeightChange={setAnnotationsHeight}
+              theme={theme}
+              xScale={xScale}
+            />
+          </g>
+        )}
+
+        <g transform={`translate(${chartStartPosition},${chartYPosition})`}>
           {reversedSeries.map((singleSeries, index) => {
             return (
               <LineSeries