+++ packages/polaris-viz/CHANGELOG.md
@@ -9,7 +9,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-- Added `annotations` to `StackedAreaChartProps`.
+- Added `annotations` to `StackedAreaChartProps` and `LineChartProps`
 
 ### Changed